refactor: keep packaged lease checks within size gates
