+++ pkg/services/factory_definitions/internal/services/distribution/packagedinstallation/diagnostics_test.go
@@ -1,12 +1,14 @@
 package packagedinstallation
 
 import (
+	"encoding/json"
 	"errors"
 	"io/fs"
 	"os"
 	"path/filepath"
 	"strings"
 	"sync"
+	"sync/atomic"
 	"testing"
 
 	platformfilesystem "github.com/portpowered/infinite-you/pkg/platform/filesystem"
@@ -175,7 +177,11 @@ func TestInstallPackagedFactory_ReclaimsOnlyRevalidatedOrphan(t *testing.T) {
 	if !foundReclaim {
 		t.Fatalf("reclaimed-orphan diagnostic missing: %#v", logger.snapshot())
 	}
-	foundAcquired := false
+	assertRecoveredAcquiredDiagnostic(t, logger)
+}
+
+func assertRecoveredAcquiredDiagnostic(t *testing.T, logger *packagedInstallationLogger) {
+	t.Helper()
 	for _, entry := range logger.snapshot() {
 		if entry.fields["outcome"] != "acquired" {
 			continue
@@ -187,9 +193,144 @@ func TestInstallPackagedFactory_ReclaimsOnlyRevalidatedOrphan(t *testing.T) {
 		if !ok || !strings.Contains(resource, "-recovered-") {
 			t.Fatalf("recovered acquisition resource = %#v, want recovery lease", entry.fields["resource"])
 		}
-		foundAcquired = true
+		return
 	}
-	if !foundAcquired {
-		t.Fatalf("recovered acquired diagnostic missing: %#v", logger.snapshot())
+	t.Fatalf("recovered acquired diagnostic missing: %#v", logger.snapshot())
+}
+
+func TestInstallPackagedFactory_ConcurrentOrphanReclaimPreservesWinnerLease(t *testing.T) {
+	root := t.TempDir()
+	name := "@test/orphan-reclaim-race"
+	stagingPath := stagingOwnershipPath(root, name)
+	if err := os.MkdirAll(stagingPath, 0o755); err != nil {
+		t.Fatalf("create orphan staging path: %v", err)
 	}
+	if err := os.WriteFile(filepath.Join(stagingPath, stagingOwnerMetadataName), []byte(`{"pid":404}`), 0o600); err != nil {
+		t.Fatalf("write orphan owner metadata: %v", err)
+	}
+	fileSystem := &orphanReclaimRaceFileSystem{
+		Local:              platformfilesystem.Local{},
+		stagingPath:        stagingPath,
+		firstReclaimReady:  make(chan struct{}),
+		secondReclaimReady: make(chan struct{}),
+		releaseFirst:       make(chan struct{}),
+		releaseSecond:      make(chan struct{}),
+	}
+	persistence := &blockingPackagedInstallationPersistence{
+		prepareStarted: make(chan struct{}),
+		allowPrepare:   make(chan struct{}),
+	}
+	params := factorydefinitions.PackagedFactoryInstallParams{
+		NamedFactoriesRoot: root,
+		BackendScopeID:     "orphan-race-scope",
+		Definition: factorydefinitions.PackagedDefinition{
+			Name: name,
+			JSON: []byte(`{}`),
+		},
+		Format: factorydefinitions.PackagedFactoryFormatJSON,
+	}
+	first := newWithOwnerProbe(
+		persistence,
+		fileSystem,
+		os.Mkdir,
+		&scriptedOwnerProbe{record: ownerRecord{PID: 101}, liveness: ownerLivenessOrphaned},
+	)
+	second := newWithOwnerProbe(
+		persistence,
+		fileSystem,
+		os.Mkdir,
+		&scriptedOwnerProbe{record: ownerRecord{PID: 202}, liveness: ownerLivenessOrphaned},
+	)
+	done := make(chan error, 2)
+	go func() {
+		_, err := first.InstallPackagedFactory(t.Context(), params)
+		done <- err
+	}()
+	go func() {
+		_, err := second.InstallPackagedFactory(t.Context(), params)
+		done <- err
+	}()
+
+	<-fileSystem.firstReclaimReady
+	<-fileSystem.secondReclaimReady
+	close(fileSystem.releaseFirst)
+	<-persistence.prepareStarted
+
+	recoveredPath := findRecoveredLeasePath(t, root)
+	ownerData, err := os.ReadFile(filepath.Join(recoveredPath, stagingOwnerMetadataName))
+	if err != nil {
+		t.Fatalf("read winner owner metadata: %v", err)
+	}
+	var winner ownerRecord
+	if err := json.Unmarshal(ownerData, &winner); err != nil {
+		t.Fatalf("decode winner owner metadata: %v", err)
+	}
+	if winner.PID != 101 && winner.PID != 202 {
+		t.Fatalf("winner owner PID = %d, want one of the contenders", winner.PID)
+	}
+	if _, err := os.Stat(stagingPath); !errors.Is(err, os.ErrNotExist) {
+		t.Fatalf("original orphan path stat = %v, want absent", err)
+	}
+
+	close(fileSystem.releaseSecond)
+	loserErr := <-done
+	if loserErr == nil || !errors.Is(loserErr, factorydefinitions.ErrFactoryInstallationContention) {
+		t.Fatalf("delayed reclaimer error = %v, want typed contention", loserErr)
+	}
+	for _, want := range []string{"outcome=indeterminate-contention", "owner_liveness=racing", stagingPath} {
+		if !strings.Contains(loserErr.Error(), want) {
+			t.Fatalf("delayed reclaimer error = %q, want %q", loserErr, want)
+		}
+	}
+	if _, err := os.Stat(filepath.Join(recoveredPath, stagingOwnerMetadataName)); err != nil {
+		t.Fatalf("winner owner metadata after delayed reclaimer = %v, want preserved", err)
+	}
+
+	close(persistence.allowPrepare)
+	if winnerErr := <-done; winnerErr != nil {
+		t.Fatalf("winning installation error = %v", winnerErr)
+	}
+	if _, err := os.Stat(recoveredPath); !errors.Is(err, os.ErrNotExist) {
+		t.Fatalf("winner recovery lease stat = %v, want released", err)
+	}
+}
+
+func findRecoveredLeasePath(t *testing.T, root string) string {
+	t.Helper()
+	entries, err := os.ReadDir(root)
+	if err != nil {
+		t.Fatalf("read root while winner is held: %v", err)
+	}
+	for _, entry := range entries {
+		if strings.Contains(entry.Name(), "-recovered-") {
+			return filepath.Join(root, entry.Name())
+		}
+	}
+	t.Fatalf("winner recovery lease was not published; entries = %v", entries)
+	return ""
+}
+
+type orphanReclaimRaceFileSystem struct {
+	platformfilesystem.Local
+	stagingPath        string
+	firstReclaimReady  chan struct{}
+	secondReclaimReady chan struct{}
+	releaseFirst       chan struct{}
+	releaseSecond      chan struct{}
+	reclaimCalls       atomic.Int32
+}
+
+func (fileSystem *orphanReclaimRaceFileSystem) Rename(oldPath, newPath string) error {
+	if oldPath != fileSystem.stagingPath {
+		return fileSystem.Local.Rename(oldPath, newPath)
+	}
+	switch fileSystem.reclaimCalls.Add(1) {
+	case 1:
+		close(fileSystem.firstReclaimReady)
+		<-fileSystem.releaseFirst
+	case 2:
+		close(fileSystem.secondReclaimReady)
+		<-fileSystem.releaseSecond
+	}
+	return fileSystem.Local.Rename(oldPath, newPath)
 }
